+++ backend/src/main/java/com/backend/member/controller/MemberController.java
@@ -6,6 +6,7 @@ import com.backend.member.dto.LoginDto;
 import com.backend.member.dto.SignupDto;
 import com.backend.member.entity.Member;
 import com.backend.member.service.MemberServiceImpl;
+import com.backend.order.dto.MemberInfoDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
@@ -124,4 +125,11 @@ public class MemberController {
         }
     }
 
+    // 멤버 정보 가져오기
+    @GetMapping("/order/getuser")
+    public MemberInfoDto getUserInfo() {
+        MemberInfoDto memberInfo = memberService.getMemberInfo();
+        System.out.println("모르곘다");
+        return memberInfo;
+    }
 }
