+++ backend/src/main/java/com/backend/order/repository/OrderRepository.java
@@ -5,5 +5,5 @@ import com.backend.order.entity.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Integer> {
-    Order findFirstByMemberOrderByOrderIdDesc(Member member);
+    Order findFirstByMemberIdOrderByOrderIdDesc(String member);
 }
