+++ backend/src/main/java/com/backend/order/service/OrderService.java
@@ -8,6 +8,4 @@ public interface OrderService {
     Integer insertOrder(OrderDTO orderDTO);
 
     AddressDTO getRecentAddr();
-
-    MemberInfoDto getMemberInfo();
 }
